--- Organization.xml
@@ -1,1 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?><organization><id>2</id><contactPhones>8(999)999-99-99</contactPhones><fullName>Организация1</fullName><head><id>0</id><firstName>Иван</firstName><lastName>Иванов</lastName><patronymic>Иванович</patronymic><position>Сотрудник</position></head><shortName>ОРГ1</shortName></organization>
