--- Storage.xml
@@ -1,1 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?><storage><docs><entry><key>0</key><value><author><id>0</id><firstName>Иван</firstName><lastName>Иванов</lastName><patronymic>Иванович</patronymic><position>Сотрудник</position></author><dateReg>2016-12-07T09:51:37.151+05:00</dateReg><id>0</id><name>Task: 0</name><regNum>D-000020</regNum><regNumCount>20</regNumCount><text>Text text text text text</text></value></entry><entry><key>1</key><value><author><id>0</id><firstName>Иван</firstName><lastName>Иванов</lastName><patronymic>Иванович</patronymic><position>Сотрудник</position></author><dateReg>2016-12-07T09:51:37.842+05:00</dateReg><id>1</id><name>Outgoing: 1</name><regNum>D-000017</regNum><regNumCount>17</regNumCount><text>Text text text text text</text></value></entry><entry><key>2</key><value><author><id>0</id><firstName>Иван</firstName><lastName>Иванов</lastName><patronymic>Иванович</patronymic><position>Сотрудник</position></author><dateReg>2016-12-07T09:51:38.192+05:00</dateReg><id>2</id><name>Incoming: 2</name><regNum>D-00008</regNum><regNumCount>8</regNumCount><text>Text text text text text</text></value></entry></docs></storage>
