+++ src/main/java/com.etorrus/Program.java
@@ -4,12 +4,12 @@ package com.etorrus;
 import com.etorrus.documents.Incoming;
 import com.etorrus.documents.Outgoing;
 import com.etorrus.documents.Task;
-import com.etorrus.factorys.*;
+import com.etorrus.factory.*;
 import com.etorrus.util.DocumentWrapper;
-import com.etorrus.service.Report;
+import com.etorrus.view.Report;
 import com.etorrus.staffs.Person;
 import com.etorrus.util.PersonWrapper;
-import com.etorrus.util.JaxbParser;
+import com.etorrus.dao.JaxbParser;
 
 import java.io.File;
 import java.util.List;
@@ -25,7 +25,7 @@ public class Program {
         System.out.println(personList.get(3).getId());
 
         JaxbParser parser = new JaxbParser();
-        File filePersons = new File("personList.xml");
+        File filePersons = new File("E:/xmlDocument/personList.xml");
 
         PersonWrapper personWrapper = new PersonWrapper();
         personWrapper.setPersonList(personList);
@@ -35,7 +35,6 @@ public class Program {
         //DocumentWrapper documentWrapper = new DocumentWrapper();
         //List<Document> documentList = new DocumentWrapper().getDocuments();
 
-        //System.out.println(((PersonWrapper) parser.XMLToStaff(filePersons, PersonWrapper.class)).getPersonList().get(3));
         List<Person> personsList = ((PersonWrapper) parser.XMLToStaff(filePersons, PersonWrapper.class)).getPersonList();
         IncomingFactory incomingFactory = new IncomingFactory();
         List<Incoming> incomingList = incomingFactory.getList(5);
@@ -48,23 +47,8 @@ public class Program {
 
 
 
-        File fileDocuments = new File("documentList.xml");
+        File fileDocuments = new File("E:/xmlDocument/documentList.xml");
         DocumentWrapper documentWrapper = new DocumentWrapper();
-
-
-
-
-
-        /*List<Person> persons = new ArrayList<Person>();
-
-        File file = new File("person.xml");
-
-        parser.staffToXML(file, person);
-        person = (Person) parser.XMLToStaff(file, Person.class);
-
-        DocumentFactory factory = new DocumentFactory();*/
-
-
 
 
         Report rep = new Report();
