+++ src/main/java/com.etorrus/documents/Storable.java
@@ -4,5 +4,4 @@ public interface Storable {
 
     public int getId();
     //public String getTable();
-
 }
