+++ src/main/java/com.etorrus/factory/DepartmentFactory.java
@@ -1,0 +1,4 @@
+package com.etorrus.factory;
+
+public class DepartmentFactory extends StaffFactory{
+}
