+++ src/main/java/com.etorrus/factory/OrganizationFactory.java
@@ -1,0 +1,4 @@
+package com.etorrus.factory;
+
+public class OrganizationFactory extends StaffFactory{
+}
