+++ src/main/java/com.etorrus/factory/StaffFactory.java
@@ -1,0 +1,9 @@
+package com.etorrus.factory;
+
+import com.etorrus.staffs.Staff;
+
+public abstract class StaffFactory {
+    public Staff create() {
+        return null;
+    }
+}
