--- src/main/java/com.etorrus/factorys/DepartmentFactory.java
@@ -1,5 +1,0 @@
-package com.etorrus.factorys;
-
-public class DepartmentFactory extends StaffFactory{
-    //jgjhgj
-}
