--- src/main/java/com.etorrus/factorys/OrganizationFactory.java
@@ -1,4 +1,0 @@
-package com.etorrus.factorys;
-
-public class OrganizationFactory extends StaffFactory{
-}
