--- src/main/java/com.etorrus/factorys/StaffFactory.java
@@ -1,9 +1,0 @@
-package com.etorrus.factorys;
-
-import com.etorrus.staffs.Staff;
-
-public abstract class StaffFactory {
-    public Staff create() {
-        return null;
-    }
-}
